Changed the main class to its own files, added the strike behaviours
