+++ PixelEngine training/8 Ball Pool X-treme/8 Ball Pool X-treme/Player.cpp
@@ -1,8 +1,10 @@
 ﻿#include "Player.h"
 
-FPlayer::FPlayer(const EBallType NewBallType, const FBall& CueBall) : Score(0), BallType(NewBallType)
+#include "EightBallPool.h"
+
+FPlayer::FPlayer(const EBallType NewBallType, const FBall* CueBall) : Score(0), BallType(NewBallType)
 {
-    Cue = new FCue(CueBall);
+    Cue = new FCue(*CueBall);
 }
 
 void FPlayer::SetScore(const int NewScore)
@@ -15,13 +17,16 @@ void FPlayer::AddScore(const int OtherScore)
     Score += OtherScore;
 }
 
-void FPlayer::Strike()
+void FPlayer::Strike(FEightBallPool* Game)
 {
     // To do: Save the initial Cue distance to the cue ball, animate the cue moving to the cue ball
     // And add the new velocity to the white ball depending on the original cue distance
+    const float OriginalDistance = Game->GetPlayer().GetCue().GetDistance();
+    olc::vf2d NewVelocity = - 5.f * OriginalDistance * Game->GetPlayer().GetCue().GetDirection();
+    Game->GetCueBall()->SetVelocity(NewVelocity);
 }
 
-const FCue& FPlayer::GetCue() const
+FCue& FPlayer::GetCue()
 {
     return *Cue;
 }
